+++ cmd/root.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -14,15 +13,13 @@ var rootCmd = &cobra.Command{
 	Short:   "Graft is a simple implementation of Raft distributed consensus",
 	Long:    `Makes any FSM distributed and resilient to single point of failure.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		// Do Stuff Here
-		// fmt.Println("run cmd")
 		cmd.Help()
 	},
 }
 
 func Execute() {
 	if err := rootCmd.Execute(); err != nil {
-		fmt.Fprintln(os.Stderr, err)
+		// fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
 }
