+++ persistent_turing.json
@@ -1,0 +1,5 @@
+{
+  "current_term": 1,
+  "voted_for": "turing",
+  "machine_logs": []
+}
