+++ pkg/domain/entity/errors.go
