+++ pkg/domain/entity/fsmState.go
