+++ pkg/domain/entity/node.go
