+++ pkg/domain/entity/role.go
