+++ pkg/domain/entity/rpc.go
