+++ pkg/domain/utils.go
